Split MoodleAPI adapter into abstract base and quiz_archiver implementation
